requirements.txt
langchain
streamlit
langchain_community
google-generativeai
langchain_google_genai